+++ src/TRegx/CleanRegex/Match/AbstractMatchPattern.php
@@ -63,7 +63,7 @@ abstract class AbstractMatchPattern implements PatternLimit, Countable
         if ($callback !== null) {
             $factory = new LazyMatchAllFactory($this->base);
             $polyfill = new GroupPolyfillDecorator($match, $factory, 0);
-            $matchObject = new MatchImpl($this->base, 0, $polyfill, $factory, $this->base->getUserData());
+            $matchObject = new MatchImpl($this->base, 0, 1, $polyfill, $factory, $this->base->getUserData());
             return $callback($matchObject);
         }
         return $match->getText();
@@ -104,7 +104,7 @@ abstract class AbstractMatchPattern implements PatternLimit, Countable
     {
         $matches = $this->base->matchAllOffsets();
         if ($matches->matched()) {
-            $result = $callback($matches->getFirstMatchObject(new MatchObjectFactoryImpl($this->base, $this->base->getUserData())));
+            $result = $callback($matches->getFirstMatchObject(new MatchObjectFactoryImpl($this->base, 1, $this->base->getUserData())));
             return new MatchedOptional($result);
         }
         return new NotMatchedOptional(
@@ -151,7 +151,7 @@ abstract class AbstractMatchPattern implements PatternLimit, Countable
      */
     protected function getMatchObjects(): array
     {
-        $factory = new MatchObjectFactoryImpl($this->base, $this->base->getUserData());
+        $factory = new MatchObjectFactoryImpl($this->base, -1, $this->base->getUserData());
         return $this->base->matchAllOffsets()->getMatchObjects($factory);
     }
 }